+++ examples/watchOS Example/watchOSsample WatchKit Extension/InterfaceController.swift
@@ -1,0 +1,31 @@
+//
+//  InterfaceController.swift
+//  watchOSsample WatchKit Extension
+//
+//  Created by Boris Bügling on 13/06/15.
+//  Copyright © 2015 Boris Bügling. All rights reserved.
+//
+
+import WatchKit
+import Foundation
+
+
+class InterfaceController: WKInterfaceController {
+
+    override func awakeWithContext(context: AnyObject?) {
+        super.awakeWithContext(context)
+        
+        // Configure interface objects here.
+    }
+
+    override func willActivate() {
+        // This method is called when watch view controller is about to be visible to user
+        super.willActivate()
+    }
+
+    override func didDeactivate() {
+        // This method is called when watch view controller is no longer visible
+        super.didDeactivate()
+    }
+
+}